refactor: split spec_cleaner to multiple sub-modules
This is a preparation for the support of IPv6 and MAC address
obfuscation.  By splitting it to multiple sub-modules, it will be easy
for adding new modules, e.g. IPv6 and MAC obfuscation.
Jira: RHINENG-15077

- replace the `get_obfuscate_functions` with a new keyword argument
  `width`, for specs that need to keep the original words width, this
  `width` needs to be set by hand before content cleaning.  In the
  future, it will be moved to the RegistryPoint for user to set by hand
  per specs.
- the processing of "password" and "keyword" is kind of "obfuscation"
  which replace the potential `password value` and specified `keywords`
  to kind of fixed strings/words, so moved it to the obfuscation module.
  User can exclude it by specifying it with, e.g.
  "no_obfuscate=['password']" while adding its RegistryPoint in
  `insights.specs.Specs`.
  Jira: RHINENG-14756
