+++ insights/cleaner/mac.py
@@ -1,0 +1,14 @@
+"""
+MAC Obfuscation
+===============
+"""
+
+import logging
+
+
+logger = logging.getLogger(__name__)
+
+
+# TODO
+class Mac(object):
+    pass
