+++ insights/tests/specs/test_specs_filters.py
@@ -6,6 +6,7 @@ from collections import defaultdict
 from mock.mock import patch
 
 from insights import collect
+from insights.cleaner import Cleaner
 from insights.client.archive import InsightsArchive
 from insights.client.config import InsightsConfig
 from insights.core import dr
@@ -13,7 +14,6 @@ from insights.core import filters
 from insights.core.context import HostContext
 from insights.core.filters import add_filter
 from insights.core.plugins import datasource
-from insights.core.spec_cleaner import Cleaner
 from insights.core import spec_factory
 from insights.core.spec_factory import (
     RegistryPoint,
@@ -311,7 +311,7 @@ def test_exp_no_filters():
 
 
 @pytest.mark.parametrize("obfuscate", [True, False])
-@patch('insights.core.spec_cleaner.Cleaner.generate_report', return_value=None)
+@patch('insights.cleaner.Cleaner.generate_report', return_value=None)
 def test_specs_filters_collect(gen, obfuscate):
     add_filter(Stuff.many_glob_filter, " ")
     add_filter(Stuff.many_foreach_exe_filter, " ")
