+++ insights/tests/specs/test_specs_special_content.py
@@ -89,7 +89,7 @@ def teardown_function(func):
 
 
 @pytest.mark.parametrize("obfuscate", [True, False])
-@patch('insights.core.spec_cleaner.Cleaner.generate_report')
+@patch('insights.cleaner.Cleaner.generate_report')
 def test_specs_special_content_collect(report, obfuscate):
     # Preparation
     manifest = collect.load_manifest(specs_manifest)
